+++ src/cellular_modem/__init__.py
@@ -2,6 +2,7 @@
 
 from .at import ATClient, ATError, ATResponse, ATTimeout, SerialTransport
 from .modem import CallInfo, Modem, SMSMessage
+from .ports import SerialPortInfo, list_serial_ports
 from .profiles import ModemProfile, get_profile
 from .smoke import SmokeCheck, SmokeReport, report_to_dict, report_to_markdown, run_read_only_smoke
 
@@ -16,8 +17,10 @@ __all__ = [
     "SMSMessage",
     "SmokeCheck",
     "SmokeReport",
+    "SerialPortInfo",
     "SerialTransport",
     "get_profile",
+    "list_serial_ports",
     "report_to_dict",
     "report_to_markdown",
     "run_read_only_smoke",